+++ mcdc_tnt/generations.py
@@ -48,14 +48,16 @@ def Generations(comp_parms, sim_perams, mesh_cap_xsec, mesh_scat_xsec, mesh_fis_
         import mcdc_tnt.pp_kernels as kernels
         
     elif comp_parms['hard_targ'] == 'nb_cpu':
+        print('loading cpu')
         import mcdc_tnt.numba_kernels.cpu as kernels
         from mcdc_tnt.numba_kernels.warmup import WarmUp
         WarmUp(comp_parms['p_warmup']) #warmup kernels
         
     elif comp_parms['hard_targ'] == 'nb_gpu':
+        print('loading gpu')
         import mcdc_tnt.numba_kernels.gpu as kernels
-        from mcdc_tnt.numba_kernels.warmup import WarmUp
-        WarmUp(comp_parms['p_warmup']) #warmup kernels
+        #from mcdc_tnt.numba_kernels.warmup import WarmUp
+        #WarmUp(comp_parms['p_warmup']) #warmup kernels
     
     N_mesh = sim_perams['N_mesh']
     nu_new_neutrons = sim_perams['nu']
