--- mcdc_tnt/numba_kernels/cuda/advance.txt
@@ -1,2 +1,0 @@
-This will be an advance kernal written for GPUs
-
