--- mcdc_tnt/numba_kernels/cuda/kernels.py
@@ -1,13 +1,0 @@
-"""
-Created on Thu Nov 18 11:46:11 2021
-
-@author: jack
-"""
-
-from mcdc.SourceParticles import SourceParticles
-import Advance.Advance as Advance
-import SampleEvent.SampleEvent as SampleEvent
-import FissionsAdd.FissionsAdd as FissionsAdd
-import CleanUp.CleanUp as BringOutYourDead
-import Scatter.Scatter as Scatter
-import SourceParticles.StillIn as StillIn
