+++ mcdc_tnt/tc_1.yaml
@@ -1,7 +1,7 @@
 
 
 name: 'fissioning_slab'
-number of particles: 1e5
+number of particles: 1e7
 rng seed: 777
 particle speed: 1
 neutrons per fission: 2
@@ -16,7 +16,7 @@ surface locations: [0,1]
 
 dx: 0.01
 
-hardware target: pyk_cpu ## pp/nb_cpu/nb_gpu/pyk_cpu/pyk_gpu
+hardware target: nb_gpu ## pp/nb_cpu/nb_gpu/pyk_cpu/pyk_gpu
 print warmup times: True
 
 assemble mesh: True #assemble mesh from crossections here or import from file
